pyproject.toml
[tool.poetry]
name = "filespeak-backend"
version = "0.1.1"
description = "Template to run fast api in fly.io with poetry as env manager"
authors = ["Giorgio Segalla <giorgiosgl1997@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
gradio = "^5.34.2"
python-docx = "^1.2.0"
pypdf2 = "^3.0.1"
python-decouple = "^3.8"
openai = "^1.88.0"
chromadb = "^1.0.13"
tiktoken = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ipython = "^9.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
